+++ pyrit/score/self_ask_category_scorer.py
@@ -80,7 +80,7 @@ class SelfAskCategoryScorer(Scorer):
 
         return category_descriptions
 
-    async def score(self, request_response: PromptRequestPiece) -> list[Score]:
+    async def score_async(self, request_response: PromptRequestPiece) -> list[Score]:
         """
         Scores the given text using the chat target.
