+++ pyrit/score/self_ask_true_false_scorer.py
@@ -66,7 +66,7 @@ class SelfAskTrueFalseScorer(Scorer):
             orchestrator_identifier=None,
         )
 
-    async def score(self, request_response: PromptRequestPiece) -> list[Score]:
+    async def score_async(self, request_response: PromptRequestPiece) -> list[Score]:
         """
         Scores the given text using the chat target.
