+++ tests/orchestrator/test_xpia_orchestrator.py
@@ -10,7 +10,7 @@ from pyrit.orchestrator import (
 )
 import pytest
 
-from unittest.mock import Mock, patch
+from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 from pyrit.prompt_target.prompt_target import PromptTarget
 from pyrit.score import Score, Scorer
@@ -34,9 +34,15 @@ def processing_target() -> PromptTarget:
 
 @pytest.fixture
 def success_scorer() -> Scorer:
-    scorer = Mock()
-    scorer.score_text = Mock(return_value=Score(score_type="bool", score_value=True))
-    return scorer
+    mock_score = MagicMock(Score)
+    mock_score.score_value = True
+    mock_score.score_type = "true_false"
+    mock_score.get_value.return_value = True
+
+    mock_scorer = MagicMock(Scorer)
+    mock_scorer.scorer_type = "true_false"
+    mock_scorer.score_text_async = AsyncMock(return_value=[mock_score])
+    return mock_scorer
 
 
 def test_xpia_orchestrator_execute_no_scorer(attack_setup_target):
@@ -70,8 +76,8 @@ def test_xpia_orchestrator_execute(attack_setup_target, success_scorer):
         processing_callback=processing_callback,
     )
     score = xpia_orchestrator.execute()
-    assert score.score_value
-    assert success_scorer.score_text.called_once
+    assert score[0].score_value
+    assert success_scorer.score_text_async.called_once
 
 
 def test_xpia_manual_processing_orchestrator_execute(attack_setup_target, success_scorer, monkeypatch):
@@ -83,8 +89,8 @@ def test_xpia_manual_processing_orchestrator_execute(attack_setup_target, succes
         scorer=success_scorer,
     )
     score = xpia_orchestrator.execute()
-    assert score.score_value
-    assert success_scorer.score_text.called_once
+    assert score[0].score_value
+    assert success_scorer.score_text_async.called_once
 
 
 def test_xpia_test_orchestrator_execute(attack_setup_target, processing_target, success_scorer):
@@ -97,8 +103,8 @@ def test_xpia_test_orchestrator_execute(attack_setup_target, processing_target, 
             scorer=success_scorer,
         )
         score = xpia_orchestrator.execute()
-        assert score.score_value
-        assert success_scorer.score_text.called_once
+        assert score[0].score_value
+        assert success_scorer.score_text_async.called_once
         assert mock_send_to_processing_target.called_once
 
 
@@ -114,7 +120,7 @@ def test_xpia_orchestrator_process_async(attack_setup_target, processing_target,
                 scorer=success_scorer,
             )
             score = xpia_orchestrator.execute()
-            assert score.score_value
-            assert success_scorer.score_text.called_once
+            assert score[0].score_value
+            assert success_scorer.score_text_async.called_once
             assert mock_send_to_processing_target.called_once
             assert mock_send_async_to_processing_target.assert_called_once
